+++ bin/parsers/__init__.py
@@ -7,16 +7,23 @@
 #  ╚═════╝╚═╝   ╚═╝   ╚═╝  ╚═╝ ╚═════╝ ╚══════╝
 # ==============================================
 
+from .run import parser_run
+from .init import parser_init
+from .doctor import parser_doctor
 from .simulation import parser_simulation
 from .parameter import parser_parameter
 from .launch import parser_launch
 from .data import parser_data
 from .report import parser_report
 
+
 __all__ = [
+    parser_run,
     parser_simulation,
     parser_parameter,
     parser_launch,
     parser_data,
     parser_report,
+    parser_init,
+    parser_doctor,
 ]
