+++ citros/parsers/__init__.py
@@ -12,6 +12,6 @@
 # from .parser_ros2 import parser_ros2
 # from .citros_integration import citros_integration
 
-from .parser_ros2 import parser_ros2
+from .parser_ros2 import ParserRos2
 
-__all__ = ["parser_ros2"]
+__all__ = ["ParserRos2"]
