+++ citros/parsers/parser_base.py
@@ -1,3 +1,3 @@
-class parser_base:
+class ParserBase:
     def parse(self):
         raise NotImplementedError()
