--- data/defaults/default_flow.json
@@ -1,25 +1,0 @@
-{
-    "trigger": "",
-    "simulations": [
-        [
-            "sim1",
-            17
-        ],
-        [
-            "sim2",
-            42
-        ],
-        [
-            "sim3",
-            69
-        ]
-    ],
-    "notebooks": [
-        "nb1.ipynb",
-        "nb2.ipynb"
-    ],
-    "recipients": [
-        "john@doe.com",
-        "enigo@montoya.com"
-    ]
-}
