--- data/defaults/default_settings.json
@@ -1,5 +1,0 @@
-{
-    "name": "default_settings",
-    "force_message": "True",
-    "force_batch_name": "True"
-}
