--- data/defaults/default_simulation.json
@@ -1,13 +1,0 @@
-{
-    "description": "Default simulation. Change the values according to your needs.",
-    "parameter_setup": "default_param_setup.json",
-    "launch": {
-        "file": "default.launch.py",
-        "package": "your.package"
-    },
-    "timeout": 60,
-    "GPU": 0,
-    "CPU": 2,
-    "MEM": 265,
-    "storage_type": "MCAP"
-}
