+++ data/doc/notebooks/README.md
@@ -1,3 +1,5 @@
+# .citros/notebooks/README.md
+
 ## Notebooks
 
 see documentation [here](https://citros.io/doc/docs_cli/structure/citros_structure#directory-notebooks)
