+++ data/doc/parameter_setups/README.md
@@ -1,3 +1,5 @@
+# .citros/parameter_setup/README.md
+
 ## Paramater Setups
 
 see documentation [here](https://citros.io/doc/docs_cli/structure/citros_structure#directory-parameter_setups)
