+++ data/doc/parameter_setups/functions/README.md
@@ -1,3 +1,5 @@
+# .citros/parameter_setups/functions
+
 ## Function Objects
 
 For a detailed description of function objects, their usage, use cases and relevant examples, please see the documentation [here](https://citros.io/doc/docs_cli/configuration/config_params).
