+++ data/doc/reports/README.md
@@ -1,3 +1,5 @@
+# .citros/reports/README.md
+
 ## Reports
 
 See documentation [here](https://citros.io/doc/docs_cli/structure/citros_structure/#directory-reports)
