+++ data/doc/simulations/README.md
@@ -1,3 +1,5 @@
+# .citros/simulation/README.md
+
 ## Simulations
 
 see documentation [here](https://citros.io/doc/docs_cli/structure/citros_structure#directory-simulations)
