--- data/markdown/README_foxglove.md
@@ -1,3 +1,0 @@
-## Foxglove Layouts
-
-Copy your Foxglove layout into this folder, thereby enabling CITROS to use them while running Foxglove on the CITROS cloud.
