--- data/markdown/README_workflows.md
@@ -1,3 +1,0 @@
-## Workflows
-
-see documentation [here](https://citros.io/doc/docs_cli/structure/citros_structure#directory-workflows)
