--- data/scripts/__init__.py
@@ -1,12 +1,0 @@
-# ==============================================
-#  ██████╗██╗████████╗██████╗  ██████╗ ███████╗
-# ██╔════╝██║╚══██╔══╝██╔══██╗██╔═══██╗██╔════╝
-# ██║     ██║   ██║   ██████╔╝██║   ██║███████╗
-# ██║     ██║   ██║   ██╔══██╗██║   ██║╚════██║
-# ╚██████╗██║   ██║   ██║  ██║╚██████╔╝███████║
-#  ╚═════╝╚═╝   ╚═╝   ╚═╝  ╚═╝ ╚═════╝ ╚══════╝
-# ==============================================
-
-# from .citros import Citros
-
-__all__ = []
